refactor(backup-agent): rename component filenames on module and lint
ref: #BKP-598
