+++ packages/manager/modules/backup-agent/src/components/CommonCells/ResourceLocationCell/__tests__/ResourceLocationCell.component.test.tsx
@@ -6,7 +6,7 @@ import { describe, expect, it, vi } from 'vitest';
 import { mockLocations } from '@/mocks/location/locations';
 import { mockVaults } from '@/mocks/vaults/vaults';
 
-import { ResourceRegionCell } from '../ResourceRegionCell.components';
+import { ResourceLocationCell } from '../ResourceLocationCell.component';
 
 const { mockLocationDetails } = vi.hoisted(() => {
   return {
@@ -24,19 +24,19 @@ vi.mock('@ovh-ux/manager-react-components', () => ({
   ),
 }));
 
-describe('ResourceRegionCell', () => {
+describe('ResourceLocationCell', () => {
   it('renders resourceName from currentState', async () => {
     mockLocationDetails.mockReturnValue({
       data: mockLocations[0],
       isLoading: false,
       isError: false,
     });
-    const vault = mockVaults[0]!;
+    const vault = { ...mockVaults[0]! };
 
-    render(<ResourceRegionCell region={vault.currentState.region} />);
+    render(<ResourceLocationCell region={vault.currentState.region} />);
 
     await waitFor(() =>
-      expect(screen.getByTestId('cell')).toHaveTextContent(mockLocations[0]!.name),
+      expect(screen.getByTestId('cell')).toHaveTextContent(mockLocations[0]!.location),
     );
     expect(mockLocationDetails).toHaveBeenCalledWith(mockVaults[0]!.currentState.region);
   });
@@ -45,7 +45,7 @@ describe('ResourceRegionCell', () => {
     mockLocationDetails.mockReturnValue({ data: null, isLoading: true, isError: false });
     const vault = mockVaults[0]!;
 
-    const { container } = render(<ResourceRegionCell region={vault.currentState.region} />);
+    const { container } = render(<ResourceLocationCell region={vault.currentState.region} />);
 
     await waitFor(() => expect(container.querySelector('ods-skeleton')).toBeVisible());
     expect(mockLocationDetails).toHaveBeenCalledWith(mockVaults[0]!.currentState.region);
@@ -55,7 +55,7 @@ describe('ResourceRegionCell', () => {
     mockLocationDetails.mockReturnValue({ data: null, isLoading: false, isError: true });
     const vault = mockVaults[0]!;
 
-    render(<ResourceRegionCell region={vault.currentState.region} />);
+    render(<ResourceLocationCell region={vault.currentState.region} />);
 
     await waitFor(() =>
       expect(screen.getByTestId('cell')).toHaveTextContent(mockVaults[0]!.currentState.region),
