+++ packages/manager/modules/backup-agent/src/components/CommonCells/ResourceStatusCell/ResourceStatusCell.component.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { DataGridTextCell } from '@ovh-ux/manager-react-components';
 
-import { ResourceStatusBadge } from '@/components/ResourceStatusBadge/ResourceStatusBadge.components';
+import { ResourceStatusBadge } from '@/components/ResourceStatusBadge/ResourceStatusBadge.component';
 import { ResourceStatus } from '@/types/Resource.type';
 
 export type ResourceStatusBadgeProps = {
