+++ packages/manager/modules/backup-agent/src/data/api/locations/location.requests.ts
@@ -1,8 +1,8 @@
-import apiClient from '@ovh-ux/manager-core-api';
+import { v2 } from '@ovh-ux/manager-core-api';
 
 import { Region } from '@/types/Location.type';
 
 const getLocationRoute = (locationName: string) => `/location/${locationName}`;
 
 export const getLocationDetails = async (locationName: string) =>
-  (await apiClient.v2.get<Region>(getLocationRoute(locationName))).data;
+  (await v2.get<Region>(getLocationRoute(locationName))).data;
