+++ packages/manager/modules/backup-agent/src/pages/services/dashboard/agents/_hooks/useAgentsListingColumns.hooks.tsx
@@ -5,7 +5,7 @@ import { useTranslation } from 'react-i18next';
 import { NAMESPACES } from '@ovh-ux/manager-common-translations';
 
 import { BACKUP_AGENT_NAMESPACES } from '@/BackupAgent.translations';
-import { ResourceStatusCell } from '@/components/CommonCells/ResourceStatusCell/ResourceStatusCell.components';
+import { ResourceStatusCell } from '@/components/CommonCells/ResourceStatusCell/ResourceStatusCell.component';
 import { AgentCreatedAtDateCell } from '@/pages/services/dashboard/agents/_components/AgentCreatedAtDateCell.component';
 import { Agent } from '@/types/Agent.type';
 import { Resource } from '@/types/Resource.type';
