+++ packages/manager/modules/backup-agent/src/pages/services/dashboard/general-information/GeneralInformation.spec.tsx
@@ -10,6 +10,7 @@ describe('[INTEGRATION] - General Information page', () => {
       initialRoute: '/vaults/dashboard/a1b2c3d4-1234-4000-82dc-5366d6786f80',
     });
 
+    /* eslint-disable max-nested-callbacks */
     await waitFor(
       () => {
         // forEach inside waitFor DOES NOT WORK
