+++ packages/manager/modules/backup-agent/src/pages/services/listing/_hooks/useVspcTenantListingColumns.tsx
@@ -3,11 +3,10 @@ import { useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 
 import { NAMESPACES } from '@ovh-ux/manager-common-translations';
-import { DatagridColumn } from '@ovh-ux/manager-react-components';
 
-import { ResourceLocationCell } from '@/components/CommonCells/ResourceLocationCell/ResourceLocationCell.components';
-import { ResourceRegionCell } from '@/components/CommonCells/ResourceRegionCell/ResourceRegionCell.components';
-import { ResourceStatusCell } from '@/components/CommonCells/ResourceStatusCell/ResourceStatusCell.components';
+import { ResourceLocationCell } from '@/components/CommonCells/ResourceLocationCell/ResourceLocationCell.component';
+import { ResourceRegionCell } from '@/components/CommonCells/ResourceRegionCell/ResourceRegionCell.component';
+import { ResourceStatusCell } from '@/components/CommonCells/ResourceStatusCell/ResourceStatusCell.component';
 import { Resource } from '@/types/Resource.type';
 import { Tenant } from '@/types/Tenant.type';
 import { WithRegion } from '@/types/Utils.type';
