+++ packages/manager/modules/backup-agent/src/pages/vaults/dashboard/buckets/VaultBuckets.page.tsx
@@ -1,6 +1,6 @@
 import React, { Suspense } from 'react';
 
-import { Link, useParams } from 'react-router-dom';
+import { useParams } from 'react-router-dom';
 
 import { useQueryClient } from '@tanstack/react-query';
 import { useTranslation } from 'react-i18next';
