+++ packages/manager/modules/backup-agent/src/pages/vaults/listing/_hooks/useColumns.hooks.tsx
@@ -1,11 +1,11 @@
 import { useTranslation } from 'react-i18next';
 
 import { BACKUP_AGENT_NAMESPACES } from '@/BackupAgent.translations';
-import { ResourceLocationCell } from '@/components/CommonCells/ResourceLocationCell/ResourceLocationCell.components';
-import { ResourceRegionCell } from '@/components/CommonCells/ResourceRegionCell/ResourceRegionCell.components';
-import { ResourceStatusCell } from '@/components/CommonCells/ResourceStatusCell/ResourceStatusCell.components';
+import { ResourceLocationCell } from '@/components/CommonCells/ResourceLocationCell/ResourceLocationCell.component';
+import { ResourceRegionCell } from '@/components/CommonCells/ResourceRegionCell/ResourceRegionCell.component';
+import { ResourceStatusCell } from '@/components/CommonCells/ResourceStatusCell/ResourceStatusCell.component';
 import { Resource } from '@/types/Resource.type';
-import { Vault, VaultResource } from '@/types/Vault.type';
+import { Vault } from '@/types/Vault.type';
 
 import { VaultBucketsCell, VaultIdCell, VaultReferenceCell } from '../_components';
 import { VaultActionCell } from '../_components/VaultActionCell.component';
