+++ packages/manager/modules/backup-agent/src/utils/__tests__/tabsWithActiveTabOrDefault.test.ts
@@ -1,7 +1,5 @@
 import { describe, expect, it } from 'vitest';
 
-import { DashboardTabType } from '@/types/Dashboard.type';
-
 import { tabsWithActiveTabOrDefault } from '../tabsWithActiveTabOrDefault';
 
 describe('tabsWithActiveTabOrDefault', () => {
